refactor(debug): Refactor debugging tools and add string processing capabilities
- Separate the original time-related test code into an independent time_test module
- Added string_test module for testing string processing functions
- Update CMakeLists.txt to include new source files
- Modify main.c to call the new test function
- Add safe string handling functions in time_tools.c
- Added string_tools.c file to implement safe string processing functions
