+++ tools-debug/string_test/string_test.h
@@ -1,0 +1,10 @@
+//
+// Created by ZetoHkr on 2025/4/12.
+//
+
+#ifndef TOOLS_DEBUG_STRING_TEST_H
+#define TOOLS_DEBUG_STRING_TEST_H
+
+int string_test(void);
+
+#endif //TOOLS_DEBUG_STRING_TEST_H
