+++ tools-debug/time_test/time_test.h
@@ -1,0 +1,10 @@
+//
+// Created by ZetoHkr on 2025/4/12.
+//
+
+#ifndef TOOLS_DEBUG_TIME_TEST_H
+#define TOOLS_DEBUG_TIME_TEST_H
+
+int time_test();
+
+#endif //TOOLS_DEBUG_TIME_TEST_H
